+++ src/aligner-fastalign/src/native/fastalign/Model.h
@@ -29,10 +29,11 @@ namespace mmt {
 
             inline void ComputeAlignments(const std::vector<std::pair<wordvec_t, wordvec_t>> &batch,
                                           std::vector<alignment_t> &outAlignments) {
-                ComputeAlignments(batch, NULL, &outAlignments);
+                ComputeAlignments(batch
+                , NULL, &outAlignments);
             }
 
-            virtual double GetProbability(word_t source, word_t target) = 0;
+            virtual float GetProbability(word_t source, word_t target) = 0;
 
             virtual void IncrementProbability(word_t source, word_t target, double amount) = 0;
